fix(hooks): resolve BigInt serialization in React Query
Fixed "Do not know how to serialize a BigInt" error in useSchedulesOptimized
by converting BigInt timestamps to numbers before passing to React Query.

Change:
- useSchedulesOptimized: startTs/endTs from BigInt to number
- Ensures React Query can serialize query keys for caching

Note:
- Proto client auto-converts number to bigint when making API calls
- JavaScript Number can safely represent Unix timestamps until year 285616

+++ web/src/hooks/useScheduleQueries.ts
@@ -51,9 +51,9 @@ export function useSchedulesOptimized(anchorDate?: Date) {
   endOfRange.setDate(now.getDate() + 15);
   endOfRange.setHours(23, 59, 59, 999);
 
-  // Convert to Unix timestamps (seconds)
-  const startTs = BigInt(Math.floor(startOfRange.getTime() / 1000));
-  const endTs = BigInt(Math.floor(endOfRange.getTime() / 1000));
+  // Convert to Unix timestamps (seconds) as numbers for React Query serialization
+  const startTs = Math.floor(startOfRange.getTime() / 1000);
+  const endTs = Math.floor(endOfRange.getTime() / 1000);
 
   return useSchedules({
     startTs,
